amvecm: Modified hist and pq_load_table struct define [2/2]
PD#173911: to compatible 64 bit with 32 bit

+++ include/linux/amlogic/media/amvecm/ve.h
@@ -61,7 +61,7 @@ struct ve_dnlp_s {
 	unsigned int white; /* 0 ~  16, weak ~ strong */
 };
 struct ve_hist_s {
-	ulong sum;
+	unsigned int sum;
 	int width;
 	int height;
 	int ave;
